pyproject.toml
[tool.poetry]
# These two lines are required
name = "lecture-navigator-server"
version = "0.2.0"

description = "FastAPI service for ingesting YouTube video transcripts and enabling semantic search with RAG."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.117.1"
uvicorn = {version = "^0.37.0", extras = ["standard"]}
youtube-transcript-api = "^1.2.2"
pydantic = "^2.7.0"
# Remove openai if not used for embeddings
# openai = "^1.51.0"
pymongo = "^4.8.0"
sentence-transformers = "^5.1.1"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"        # Add python-dotenv
llama-index-core = "^0.14.2"
## Removed: using direct OpenRouter/Gemini client calls instead of LlamaIndex LLMs
# llama-index-llms-google = "^0.3.6"
llama-index-vector-stores-mongodb = "^0.8.1"
llama-index-readers-youtube-transcript = "^0.4.1"
loguru = "^0.7.2"
python-multipart = "^0.0.9"      # For FastAPI file uploads
rerankers = "^0.5.0"             # Optional: extra re-ranker models support
requests = "^2.32.0"             # For evaluation script
pandas = "^2.2.0"                # For evaluation script
chromadb = "^0.5.5"
# Skip FAISS on Windows via marker; install manually (or use WSL/conda) if needed on Windows
faiss-cpu = { version = "1.8.0.post1", markers = "platform_system != 'Windows'" }
llama-index-vector-stores-pinecone = "^0.4.2"
pinecone-client = "^5.0.0"
llama-index-embeddings-huggingface = "^0.3.1"